requirements.txt
streamlit==1.29.0
langchain==0.1.20
langchain-openai==0.1.0
langchain-community==0.0.38
openai>=1.10.0,<2.0.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
pandas
pydantic==2.5.3
lxml==5.0.0
html5lib==1.1